fix: notifications service split
