--- src/store/auth/index.ts
@@ -1,3 +1,0 @@
-export * from './actions'
-export * from './reducer'
-export { default as authSaga } from './saga'
