+++ src/store/configureSaga.ts
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects'
-import { authSaga } from './auth'
-import { uiSaga } from './ui'
+import authSaga from './auth/saga'
+import uiSaga from './ui/saga'
 
 export default function* rootSaga() {
   yield all([...authSaga, ...uiSaga])
