--- src/store/ui/index.ts
@@ -1,3 +1,0 @@
-export * from './actions'
-export * from './reducer'
-export { default as uiSaga } from './saga'
